requirements.txt
django

djangorestframework
django-filter
markdown
djangorestframework-simplejwt

django-ninja
django-ninja-extra
orjson

python-dotenv

loguru

requests
aiohttp
bs4

nicegui
